Add files via upload
In this project, I build a Convolutional Neural Network model using TensorFlow to recognize digits. 
A convolutional neural network (CNN, or ConvNet) is a Deep Learning algorithm that can take in an input image, assign importance (learnable weights and biases) to various aspects/objects in the image, and be able to differentiate one from the other.